requirements.txt
streamlit>=1.35
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
openai>=1.44
numpy>=1.26
pytest>=8.2
pyflakes>=3.2
